+++ config.dist.php
@@ -47,6 +47,21 @@ const TITLE = 'My oPodSync server';
 const DISABLE_USER_METADATA_UPDATE = false;
 
 /**
+ * KARADAV_URL
+ * Set to the URL of an external KaraDAV server if you want to use this oPodSync instance
+ * as an external app inside KaraDAV.
+ *
+ * If enabled, an account will be created in oPodSync for any KaraDAV user
+ * if they don't have one already.
+ *
+ * Don't forget the API key in the URL.
+ *
+ * Default: null (disabled)
+ * @var string|null
+ */
+//const KARADAV_URL = 'https://:abcd@karadav.example.org/';
+
+/**
  * DATA_ROOT
  * Path of directory where data will be stored.
  * ROOT is the root of oPodSync code.
@@ -74,7 +89,7 @@ const DB_FILE = DATA_ROOT . '/data.sqlite';
 /**
  * SQLITE_JOURNAL_MODE
  * SQLite3 journaling mode
- * Default: TRUNCATE (slower)
+ * Default: TRUNCATE (slower, but safer)
  * Recommended: WAL (faster, but read below)
  *
  * If your database file is on a local disk, you will get better performance by using