+++ server/lib/OPodSync/API.php
@@ -362,10 +362,9 @@ class API
 
 	public function getRequestURI(): string
 	{
-		$url = '/' . trim($_SERVER['REQUEST_URI'] ?? '', '/');
+		$url = parse_url($_SERVER['REQUEST_URI'] ?? '', PHP_URL_PATH);
+		$url = '/' . trim($url, '/');
 		$url = substr($url, strlen($this->base_path));
-		$url = strtok($url, '?');
-		strtok('');
 		return $url;
 	}