+++ server/sql/migration_20251211.sql
@@ -1,0 +1,5 @@
+ALTER TABLE users ADD COLUMN external_user_id INTEGER NULL;
+
+DROP INDEX users_name;
+CREATE UNIQUE INDEX users_unique ON users (name, external_user_id);
+